e2e-benchmark/requirements.txt
openai>=1.30.0
anthropic>=0.25.0
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
rich>=13.7.0
typer>=0.12.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
aiohttp>=3.9.0
numpy>=1.26.0
tabulate>=0.9.0
